Simplify test config types
