+++ crates/private/support/sel4-simple-task/config-types/Cargo.nix
@@ -4,15 +4,12 @@
 # SPDX-License-Identifier: BSD-2-Clause
 #
 
-{ mk, localCrates, versions, serdeWith }:
+{ mk, localCrates, serdeWith }:
 
 mk {
   package.name = "sel4-simple-task-config-types";
   dependencies = {
-    inherit (versions) cfg-if;
     serde = serdeWith [ "derive" ];
-  };
-  target."cfg(target_os = \"none\")".dependencies = {
     inherit (localCrates)
       sel4
       sel4-simple-task-threading
