+++ crates/private/support/sel4-simple-task/config-types/src/lib.rs
@@ -8,50 +8,49 @@
 
 use core::marker::PhantomData;
 
+use sel4_simple_task_threading::StaticThread;
 use serde::{Deserialize, Serialize};
-
-cfg_if::cfg_if! {
-    if #[cfg(target_os = "none")] {
-        mod when_sel4;
-        pub use when_sel4::*;
-    } else {
-        mod when_not_sel4;
-        pub use when_not_sel4::*;
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Serialize, Deserialize)]
-#[serde(transparent)]
-pub struct ConfigBadge(RawConfigWord);
-
-impl ConfigBadge {
-    pub fn new(word: RawConfigWord) -> Self {
-        Self(word)
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Serialize, Deserialize)]
-#[serde(transparent)]
-pub struct ConfigCPtrBits(RawConfigWord);
-
-impl ConfigCPtrBits {
-    pub fn new(word: RawConfigWord) -> Self {
-        Self(word)
-    }
-}
 
 #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Serialize, Deserialize)]
 #[serde(transparent)]
 pub struct ConfigCPtr<T> {
     phantom: PhantomData<T>,
-    bits: ConfigCPtrBits,
+    bits: sel4::CPtrBits,
 }
 
 impl<T> ConfigCPtr<T> {
-    pub fn new(bits: ConfigCPtrBits) -> Self {
+    pub fn new(bits: sel4::CPtrBits) -> Self {
         Self {
             phantom: PhantomData,
             bits,
         }
+    }
+}
+
+impl<T: WrappedCPtr> ConfigCPtr<T> {
+    pub fn get(&self) -> T {
+        T::wrap(self.bits)
+    }
+}
+
+pub trait WrappedCPtr {
+    fn wrap(bits: sel4::CPtrBits) -> Self;
+}
+
+impl WrappedCPtr for sel4::CPtr {
+    fn wrap(bits: sel4::CPtrBits) -> Self {
+        Self::from_bits(bits)
+    }
+}
+
+impl<T: sel4::CapType> WrappedCPtr for sel4::Cap<T> {
+    fn wrap(bits: sel4::CPtrBits) -> Self {
+        Self::from_bits(bits)
+    }
+}
+
+impl WrappedCPtr for StaticThread {
+    fn wrap(bits: sel4::CPtrBits) -> Self {
+        Self::new(sel4::cap::Endpoint::from_bits(bits))
     }
 }
