+++ crates/private/tests/capdl/threads/Cargo.nix
@@ -14,6 +14,7 @@ mk {
       sel4
       sel4-sync
       sel4-simple-task-runtime
+      sel4-simple-task-threading
       sel4-simple-task-config-types
       sel4-test-capdl
     ;
