+++ crates/private/tests/capdl/threads/Cargo.toml
@@ -20,6 +20,7 @@ license = "BSD-2-Clause"
 sel4 = { path = "../../../../sel4" }
 sel4-simple-task-config-types = { path = "../../../support/sel4-simple-task/config-types" }
 sel4-simple-task-runtime = { path = "../../../support/sel4-simple-task/runtime" }
+sel4-simple-task-threading = { path = "../../../support/sel4-simple-task/threading" }
 sel4-sync = { path = "../../../../sel4-sync" }
 sel4-test-capdl = { path = "../../../support/sel4-test-capdl" }
 serde = { version = "1.0.228", default-features = false, features = ["alloc", "derive"] }
